+++ src/content/mod.rs
@@ -24,19 +24,25 @@ pub enum Error {
     UnexpectedDataType,
     #[cfg(feature = "_cargo_insta_internal")]
     MissingField,
+    #[cfg(feature = "_cargo_insta_internal")]
+    FileIo(std::io::Error, std::path::PathBuf),
 }
 
 impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
-            Self::FailedParsingYaml(p) => {
+            Error::FailedParsingYaml(p) => {
                 f.write_str(format!("Failed parsing the YAML from {:?}", p.display()).as_str())
             }
-            Self::UnexpectedDataType => {
+            Error::UnexpectedDataType => {
                 f.write_str("The present data type wasn't what was expected")
             }
             #[cfg(feature = "_cargo_insta_internal")]
-            Self::MissingField => f.write_str("A required field was missing"),
+            Error::MissingField => f.write_str("A required field was missing"),
+            #[cfg(feature = "_cargo_insta_internal")]
+            Error::FileIo(e, p) => {
+                f.write_str(format!("File error for {:?}: {}", p.display(), e).as_str())
+            }
         }
     }
 }