+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 All notable changes to insta and cargo-insta are documented here.
 
+## 1.37.0
+
+- All macros for file snapshots should now handle trailing commas (but not yet inline snapshots)
+
+- Vendored old `yaml-rust` dependency to avoid rustsec warnings.  #465
+
+## 1.36.1
+
+- Fix an ownership issue introduced in 1.36 with snapshot assertions.  #453
+
+## 1.36.0
+
+- Deprecate `INSTA_FORCE_UPDATE_SNAPSHOTS` env-var for `INSTA_FORCE_UPDATE`.
+  The latter was documented, the former was implemented.  #449
+
+- Add `require_full_match` option.  #448 
+
+- Deprecate `assert_display_snapshot!`.  #385
+
 ## 1.35.1
 
 - Fixed a bug with diffs showing bogus newlines.